Add a method to create the index files given a fasta file.
Addresses #4.

For this to work I needed encapsulate a call to _bwa_index(argc, argv[]). This method are its dependencies are not available in the libbwa produced by
our current bwa commit dependency. A small change (sed command) in src/main/c/Makefile alters bwa/Makefile (after download) to include all .o in the libbwa.a; perhaps
just a subset is needed but it is not clear what subset it is.