Now one can specify the expected insert size distribution as supposed… (#19)
* Now one can specify the expected insert size distribution as supposed to allow
BWA to infer it on the fly. This might have some value when re-aligning a small number of reads (inference may fail)
or when one knows that there is a heterozygous variant and we actually now the expected insert size distribution; the
input data would have a mixture of gaussians rather than the expected single gaussian.